Add style to checkout page

+++ cart/contexts.py
@@ -24,7 +24,7 @@ def cart_contents(request):
         else:
             menu_item = get_object_or_404(MenuItem, pk=item_id)
             for size, quantity in item_data['items_by_size'].items():
-                total += quantity * float(size)
+                total += quantity * int(size)
                 menu_item_count += quantity
                 cart_items.append({
                     'item_id': item_id,
